# feat: optimise AND queries