Changes the way that ANDs are converted to SQL. This now uses less subqueries which has lead to a search time decrease when searching in TS.

My benchmark search previously took ~1min in TS and roughly ~46s when running the generated SQL to an external tool.
With these improvements the benchmark search now takes ~36s in TS and roughly ~1s in the external tool. This means there are further performance gains possible by optimising other parts of the search code.

Note:
- Currently the tag ids are hardcoded into the main query; I also tried directly querying these in the main query instead, but this only lead to a ~6s improvement while breaking some tests so I will not pursue this for now (should someone be interested in how I did that, checkout this commit 2615e7d from #673 that was reverted there)