@Computerdores 2025-01-02
This should improve search times by another 50% in TS.
I am not sure why this was necessary, but I achieved it by only counting on the `id` column instead of on the full data that will be returned by `search_library`. In an external tool this only makes a difference of 0.2s out of 1.3s, but in TS this reduces the execution time of the counting query from ~16s to ~1s for my benchmark example.