@Computerdores 2025-01-02
I think if the `search_library` method only returned the ids of the Entries this would probably cut the search time for my benchmark example from the remaining ~17s down to ~2s, but this would be a larger endeavor since it would need substantial changes though out the project
@CyanVoxel If I understand correctly what you are doing in #655 atm, then you are moving parts of the Qt frontend towards only keeping the Entry ids around and not the full entries - do you think that would make this change more feasible?