@CyanVoxel 2025-01-02
> I think if the `search_library` method only returned the ids of the Entries this would probably cut the search time for my benchmark example from the remaining ~36s down to ~17s, but this would be a larger endeavor since it would need substantial changes though out the code base. @CyanVoxel If I understand what you are doing in #655 atm correctly, then you are moving parts of the Qt frontend towards only keeping the Entry ids around and not the full entries - do you think that would make this change more feasible?

As of my latest commit in #655 (f59b84b5294b171bb6debfa3ffd2c8942e68cb73) I *believe* that this would be okay, as the main call for `search_library()` now immediately converts those results to a list of ID ints immediately anyway. The only other call is in `dupe_files.py` which I think can be refactored to work with this. So if returning IDs offers a performance improvement (and is THAT big) I think we should do it